Split off Phonetic Base into its own thing
