+++ Ursus Sigil/Sigil.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import SwiftSVG
 
 public struct Sigil {
